% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness_scores.R
\name{keyness_scores}
\alias{keyness_scores}
\title{Calculate observed keyness scores}
\usage{
keyness_scores(ifl, type = "llr", laplace = 1)
}
\arguments{
\item{ifl}{Indexed frequency list as generated by \code{create_ifl()}.}

\item{type}{The type of keyness measure. One of \code{llr}, \code{chisq}, \code{diff} or \code{logratio}. See details.}

\item{laplace}{Parameter of laplace correction. Only relevant for \code{type = "logratio"}. See details.}
}
\value{
a numerical vector of the scores, one for each term.
}
\description{
Calculates a vector of observed keyness scores for a given pair of corpora.
}
\details{
Keyness scores are calculated for an Indexed frequency list from a given pair of corpora 
as generated by \code{create_ifl()}.

Currently, the following types of scores are supported:
\describe{
    \item{\code{llr}}{The log-likelihood ratio}
    \item{\code{chisq}}{The Chi-Square-Statistic}
    \item{\code{diff}}{Difference of relative frequencies}
    \item{\code{logratio}}{Binary logarithm of the ratio of the relative frequencies, possibly using a laplace correction to avoid infinite values.}
 }
}
