% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyperm-package.R
\docType{package}
\name{keyperm-package}
\alias{keyperm}
\alias{keyperm-package}
\title{keyperm: Keyword Analysis Using Permutation Tests}
\description{
Implementation of permutation-based keyword analysis for corpus linguistics.
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()]. 
Otherwise refer to the vignettes to see how to format the documentation.
}
\author{
\strong{Maintainer}: Thoralf Mildenberger \email{mild@zhaw.ch} (\href{https://orcid.org/0000-0001-7242-1873}{ORCID})

}
\keyword{internal}
