% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyperm.R
\name{keyperm}
\alias{keyperm}
\title{Calculate the permutation distribution for a keyness measure}
\usage{
keyperm(ifl, observed, type = "llr", laplace = 1, output = 2, nperm)
}
\arguments{
\item{ifl}{Indexed frequency list as generated by \code{create_ifl()}.}

\item{observed}{The vector of observed values of the keyness scores as generarted by \code{keyness_scores()}}

\item{type}{The type of keyness measure. One of \code{llr}, \code{chisq}, \code{diff} or \code{logratio}. See details.}

\item{laplace}{Parameter of laplace correction. Only relevant for \code{type = "logratio"}. See details.}

\item{output}{The type of output. \code{output = 1} a matrix with all generated scores is returned,
for \code{output = 2} a matrix with three columns counting the number of permutations for which the 
score is strictly smaller than, equal to or strictly larger than the observed value.}

\item{nperm}{The number of permutations to generate.}
}
\value{
A numeric matrix with number of rows equal to the number of terms. The columns contain either all permutation values
    of the keyness score (\code{output = 1}) or the number of permutations for which the 
    score is strictly smaller than, equal to or strictly larger than the observed value (\code{output = 2}).
}
\description{
Calculate the permutation distributions of a given keyness measure for each term by 
shuffeling the copus labels. Number of documents per corpus is kept constant.
}
\details{
Currently, the following types of scores are supported:
\describe{
    \item{\code{llr}}{The log-likelihood ratio}
    \item{\code{chisq}}{The Chi-Square-Statistic}
    \item{\code{diff}}{Difference of relative frequencies}
    \item{\code{logratio}}{Binary logarithm of the ratio of the relative frequencies, possibly using a laplace correction to avoid infinite values.}
 }
}
